+++ src/server/game/Entities/Player/Player.cpp
@@ -2818,7 +2818,7 @@ void Player::InitStatsForLevel(bool reapplyMods)
     InitDataForForm(reapplyMods);
 
     // save new stats
-    for (uint8 i = POWER_MANA; i < MAX_POWERS; ++i)
+    for (uint16 i = POWER_MANA; i < MAX_POWERS; ++i)
         SetMaxPower(Powers(i), uint32(GetCreatePowerValue(Powers(i))));
 
     SetMaxHealth(classInfo.basehealth);                     // stamina bonus will applied later
